frontback integrated/test/requirements.txt
----------------------------- ---------
bcrypt 4.2.1
channels 4.2.0
composer 0.28.0
django-celery-beat 2.7.0
django-celery-results 2.5.1
django-cors-headers 4.4.0
django-filter 24.3
django_limiter 1.0.2
django-ratelimit 4.1.0
djangorestframework-simplejwt 5.3.1
djongo 1.3.6
easyocr 1.7.2
esbonio 0.16.5
face-recognition 1.3.0
facenet-pytorch 2.6.0
fuzzywuzzy 0.18.0
imutils 0.5.4
keras-models 0.0.7
node 1.2.2
npm 0.1.1
npx 0.1.6
nvm 1.0.21
ocrmypdf 16.8.0
opencv-contrib-python 4.10.0.84
pdf2image 1.17.0
php 1.2.1
pip 24.3.1
PyQt5 5.15.11
pytesseract 0.3.13
python-decouple 3.8
python-docx 1.1.2
python-dotenv 1.0.1
python-Levenshtein 0.26.1
redis 5.2.0
reportlab 4.2.5
rest-framework-simplejwt 0.0.2
sentencepiece 0.2.0
simplejwt 2.0.1
tensorflow 2.18.0
tesserocr 2.7.1
torchaudio 2.2.2
transformers 4.48.3
ultralytics 8.3.50
vosk 0.3.44
yolo 0.3.2